Add backend validation and fix bug related to order/start ride

+++ backend/GetGo-BE/src/main/java/rs/getgo/backend/dtos/driver/GetDriverDTO.java
@@ -18,7 +18,7 @@ public class GetDriverDTO {
     private String phone;
     private String address;
     private String profilePictureUrl;
-    private Integer recentHoursWorked;
+    private Double recentHoursWorked;
     // Vehicle info
     private String vehicleModel;
     private String vehicleType;
